added registration1

+++ src/main/java/com/ships/Controllers/AdminWindowController.java
@@ -8,10 +8,7 @@ import javafx.collections.FXCollections;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.scene.Node;
-import javafx.scene.control.Button;
-import javafx.scene.control.SelectionMode;
-import javafx.scene.control.TableColumn;
-import javafx.scene.control.TableView;
+import javafx.scene.control.*;
 import javafx.scene.control.cell.CheckBoxTableCell;
 import javafx.scene.control.cell.ComboBoxTableCell;
 import javafx.scene.control.cell.PropertyValueFactory;
@@ -21,6 +18,8 @@ import org.hibernate.ReplicationMode;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.hibernate.cfg.Configuration;
+
+import java.util.Optional;
 
 /**
  * Created by admin on 13.04.17.
@@ -48,32 +47,26 @@ public class AdminWindowController {
     private Button btnOk;
 
     public void initialize(){
+
+    }
+
+    public void fillTableView() {
+
         employeeList.setCollection(FXCollections.observableArrayList(session.createCriteria(Employee.class).list()));
         tableViewAcc.getSelectionModel().setSelectionMode(SelectionMode.SINGLE);
         tableColumnCode.setCellValueFactory(new PropertyValueFactory<Employee, String>("login"));
         tableColumnEmail.setCellValueFactory(new PropertyValueFactory<Employee, String>("email"));
-       tableColumnActive.setCellValueFactory(new PropertyValueFactory<Employee, Boolean>("active"));
-//        tableColumnActive.setCellValueFactory(new Callback<TableColumn.CellDataFeatures<Employee,Boolean>,ObservableValue<Boolean>>()
-//        {
-//            //This callback tell the cell how to bind the data model 'Registered' property to
-//            //the cell, itself.
-//           // @Override
-//            public ObservableValue<Boolean> call(TableColumn.CellDataFeatures<Employee, Boolean> param)
-//            {
-//                //param = new SimpleBooleanProperty(param.getValue().getActive());
-//                return param.getValue().activeProperty();
-//            }
-//        });
-        //tableColumnActive.setCellFactory(CheckBoxTableCell.forTableColumn(tableColumnActive));
+        tableColumnActive.setCellValueFactory(new PropertyValueFactory<Employee, Boolean>("active"));
         tableColumnActive.setCellFactory(tableColumnActive -> new CheckBoxTableCell());
         tableColumnActive.setEditable(true);
         tableViewAcc.setEditable(true);
         tableViewAcc.setItems(employeeList.getCollection());
-        employeeList.printCollection();
+
     }
+
     @FXML
     void onBtnAction(ActionEvent event) {
-        //employeeList.printCollection();
+
         Object source = event.getSource();
         if(!(source instanceof Button))
         {return;}
@@ -81,11 +74,14 @@ public class AdminWindowController {
         Button clicked = (Button) source;
 
         Window parent = ((Node)event.getSource()).getScene().getWindow();
-        // clicked.setEffect(new DropShadow());
 
         switch (clicked.getId()){
 
             case "btnOk":{
+                if(session.isOpen()!=true){
+                    session=sessionFactory.openSession();
+                }
+                session = sessionFactory.openSession();
                 session.beginTransaction();
                 employeeList.printCollection();
                 for (Employee j:employeeList.getCollection()) {
@@ -93,15 +89,25 @@ public class AdminWindowController {
                     session.getTransaction().commit();
                 }
 
-                session.close();
+               session.close();
             break;
             }
             case "btnDelete":{
-                session.beginTransaction();
-                session.delete((Employee)tableViewAcc.getSelectionModel().getSelectedItem());
-                employeeList.delete((Employee)tableViewAcc.getSelectionModel().getSelectedItem());
-                session.getTransaction().commit();
-                session.close();
+                if(session.isOpen()!=true) {
+                    session = sessionFactory.openSession();
+                }
+                Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+                alert.setTitle("Confirm");
+                alert.setHeaderText(null);
+                alert.setContentText("Do you want to delete selected user?");
+                Optional<ButtonType> result = alert.showAndWait();
+                if(result.get()==ButtonType.OK) {
+                    session.beginTransaction();
+                    session.delete((Employee) tableViewAcc.getSelectionModel().getSelectedItem());
+                    employeeList.delete((Employee) tableViewAcc.getSelectionModel().getSelectedItem());
+                    session.getTransaction().commit();
+                }
+               session.close();
                 break;
             }
         }
